--- a/frappe/utils/__init__.py
+++ b/frappe/utils/__init__.py
@@ -9,7 +9,6 @@ import os, re, sys, json, hashlib, traceback
 import functools
 from .html_utils import sanitize_html
 import frappe
-from frappe.utils.identicon import Identicon
 from email.utils import parseaddr, formataddr
 from email.header import decode_header, make_header
 # utility functions like cint, int, flt, etc.
@@ -201,6 +200,8 @@ def get_gravatar_url(email):
 	return "https://secure.gravatar.com/avatar/{hash}?d=mm&s=200".format(hash=hashlib.md5(email.encode('utf-8')).hexdigest())
 
 def get_gravatar(email):
+	from frappe.utils.identicon import Identicon
+
 	gravatar_url = has_gravatar(email)
 
 	if not gravatar_url: