--- a/frappe/utils/html_utils.py
+++ b/frappe/utils/html_utils.py
@@ -5,7 +5,6 @@ import re
 import bleach
 import bleach_whitelist.bleach_whitelist as bleach_whitelist
 from six import string_types
-from bs4 import BeautifulSoup
 
 def clean_html(html):
 	if not isinstance(html, string_types):
@@ -41,6 +40,8 @@ def clean_email_html(html):
 
 def clean_script_and_style(html):
 	# remove script and style
+	from bs4 import BeautifulSoup
+
 	soup = BeautifulSoup(html, 'html5lib')
 	for s in soup(['script', 'style']):
 		s.decompose()
@@ -53,6 +54,8 @@ def sanitize_html(html, linkify=False):
 
 	Does not sanitize JSON, as it could lead to future problems
 	"""
+	from bs4 import BeautifulSoup
+
 	if not isinstance(html, string_types):
 		return html