This doesn't look tasteful, but at least it's fast. Well, I'll let you decide if it's fast. This is how long a typical bench CLI invocation takes.
```
$ time bench --site speed.com execute frappe.ping  
1.03s user 0.13s system 100% cpu 1.152 total
```
A lot of this isn't frappe's fault. e.g. Invoking the same command without bench.
```
cd sites
time python -m frappe.utils.bench_helper frappe --site speed.com execute frappe.ping
0.62s user 0.04s system 99% cpu 0.659 total
```
We can improve this by not importing code that isn't executed e.g. Jinja, Faker, Markdown2, etc, and shed ~400ms.
```
$ time python -m frappe.utils.bench_helper frappe --site speed.com execute frappe.ping
0.20s user 0.03s system 99% cpu 0.230 total
```
### Profiling
Python provides a non-invasive way to profile import times with [`-X importtime`](https://docs.python.org/3/using/cmdline.html#id5). 
```
python -X importtime -m frappe.utils.bench_helper frappe --site speed.com execute frappe.ping
```
and get
```
import time: self [us] | cumulative | imported package
import time:       101 |        101 | zipimport
import time:       630 |        630 | _frozen_importlib_external
...
...
import time:       240 |        396 |   frappe.core.doctype.access_log.access_log
import time:       314 |       1011 | frappe.utils.response
"pong"
```
We can go through this like a caveman or we can use [tuna](https://github.com/nschloe/tuna) for visualization with something like.
```
python -X importtime -m frappe.utils.bench_helper frappe --site speed.com execute frappe.ping 2> import.log
tuna import.log
```
### Screenshots
**Before**
![Screenshot_2021-02-19 tuna - import log(1)](https://user-images.githubusercontent.com/8528887/108663937-b194df80-74f7-11eb-8199-42f5a6112864.png)

**After**
![Screenshot_2021-02-22 tuna - import log(1)](https://user-images.githubusercontent.com/8528887/108664492-e2294900-74f8-11eb-9adc-d894b5edf503.png)

https://github.com/frappe/bench/pull/1124 fixes Bench's issue.

Before Both PRs
```
$ time bench --site speed.com execute frappe.ping  
1.03s user 0.13s system 100% cpu 1.152 total
```
Before this PR (and after Bench PR)
```
$ time bench --site speed.com execute frappe.ping  
0.69s user 0.09s system 100% cpu 0.770 total
```
After both PRs
```
$ time bench --site speed.com execute frappe.ping  
0.28s user 0.06s system 101% cpu 0.329 total
```

### Possible Improvement
Werkzeug (Imported for werkzeug.Local) adds another ~100ms, maybe we can get rid of that as well.