@leela 2021-02-22
PEP 8 suggests to put imports always on top of the file (https://www.python.org/dev/peps/pep-0008/#imports). I agree that you can get performance benefit with import's inside functions for CLI's but not really for web requests. 

Here are few thoughts about having import's inside a function
1. import duplication, we end up having same import in multiple functions in the same file.
2. First import is expensive but importing the same module again is cheap (so, doesn't really benefit us in web applications)
3. Having all import's on top gives us high level idea about modules's dependencies.
4. Circular dependancies always tells us that code can be organised better and make us to push in that direction. So, if there are any import's inside a function that is sign of needed refactoring.

I agree, at some places we can't avoid circular imports and and in those cases I think we should go for import's inside functions.